+++ App.tsx
@@ -2,7 +2,7 @@ import { View, Text, StatusBar, TurboModuleRegistry } from 'react-native'
 import React from 'react'
 
 import Notifications from './src/screens/notifications/Notifications'
-import DActiveExercises from './src/screens/d-active exercises/DActiveExercises'
+import DActiveExercises from './src/screens/dActive exercises/DActiveExercises'
 
 const App = () => {
   const showNotificationScreen = true
