+++ src/constants/dactive-constants.ts
@@ -1,24 +1,24 @@
-import {ASDActiveCardStyles} from '../components/d-active-card/asDActiveCard-styles';
+import { ASDActiveCardStyles } from '../components/dActive-card/asDActiveCard-styles'
 
 export const DActiveCardInfo = {
   'outer ring': {
     imageSrc: require('../assets/images/outer-ring.png'),
     textColour: ASDActiveCardStyles.outerRingText,
   },
-  focus: {
+  'focus': {
     imageSrc: require(`../assets/images/focus.png`),
     textColour: ASDActiveCardStyles.focusText,
   },
-  follow: {
+  'follow': {
     imageSrc: require(`../assets/images/follow.png`),
     textColour: ASDActiveCardStyles.followText,
   },
-  scan: {
+  'scan': {
     imageSrc: require(`../assets/images/scan.png`),
     textColour: ASDActiveCardStyles.scanText,
   },
-  square: {
+  'square': {
     imageSrc: require(`../assets/images/square.png`),
     textColour: ASDActiveCardStyles.squareText,
   },
-};
+}
